+++ src/core/index.ts
@@ -1,4 +1,4 @@
 export { Img, setDefaultOptixFlowConfig } from "./Img.js";
 export type { ImgProps } from "./Img.js";
-export { OptixFlowConfig } from "./OptixFlowConfig.js";
-export type { OptixFlowConfigProps } from "./OptixFlowConfig.js";
+export { ImgDefaults } from "./ImgDefaults.js";
+export type { ImgDefaultsProps } from "./ImgDefaults.js";